backend/userauths/views.py
from django.shortcuts import render
from rest_framework_simplejwt.views import TokenObtainPairView
from userauths.serializers import MyTokenObtainPairSerializer, RegisterSerializer, UserSerializer, ProfileSerializer

# Create your views here.
class MyTokenObtainPairView(TokenObtainPairView):
    serializer_class = MyTokenObtainPairSerializer