backend/requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.7
django-import-export==4.3.7
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
PyJWT==2.9.0
shortuuid==1.0.13
sqlparse==0.5.3
tablib==3.8.0
